backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.9
python-dotenv==1.0.1
slowapi==0.1.9
loguru==0.7.2
pytest==8.3.2
pytest-asyncio==0.23.8
PyJWT==2.9.0
groq==0.31.0
redis==6.4.0
tenacity==8.2.2
httpx==0.28.1